refactored aim - moved the code to its own file
