+++ examples/slurm_orion/config.yaml
@@ -8,6 +8,12 @@ exp_name: my_exp_1
 # results will not be reproducible)
 seed: 1234
 
+experiment_loggers:
+  tensorboard: null  # no parameters for tensorboard
+  aim:
+    # change this to an absolute path to always use the same aim db file
+    log_folder: ./
+
 # architecture
 num_classes: 10
 architecture: simple_mlp
